Decouple form and filters parameters
